+++ src/lib/const.ts
@@ -485,8 +485,8 @@ export const SEPARATOR = `\n\n${color.yellow(
   `^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^`,
 )}\n\n`;
 
-const RUN_DEPLOYED_SERVICES_FUNCTION = "runDeployedServices";
-export const RUN_DEPLOYED_SERVICES_FUNCTION_CALL = `${RUN_DEPLOYED_SERVICES_FUNCTION}()`;
+export const RUN_DEPLOYED_SERVICES_FUNCTION_NAME = "runDeployedServices";
+export const RUN_DEPLOYED_SERVICES_FUNCTION_CALL = `${RUN_DEPLOYED_SERVICES_FUNCTION_NAME}()`;
 
 const RUN_DEPLOYED_SERVICE_AQUA = `
 -- example of running services deployed using \`${CLI_NAME} deal deploy\`
@@ -498,7 +498,7 @@ data Answer:
     answer: ?string
     worker: Worker
 
-func ${RUN_DEPLOYED_SERVICES_FUNCTION}() -> []Answer:
+func ${RUN_DEPLOYED_SERVICES_FUNCTION_NAME}() -> []Answer:
     deals <- Deals.get()
     dealId = deals.${DEFAULT_DEAL_NAME}!.dealIdOriginal
     answers: *Answer
