+++ test/tests/deal/dealDeploy.test.ts
@@ -17,18 +17,19 @@
 import assert from "node:assert";
 import { join, relative } from "node:path";
 
-import { initServiceConfig } from "../../src/lib/configs/project/service.js";
-import { MY_SERVICE_NAME, NEW_SPELL_NAME } from "../constants.js";
-import { fluence, initializeTemplate } from "../helpers/common.js";
-import { maybeConcurrentTest } from "../helpers/testWrapper.js";
+import { initServiceConfig } from "../../../src/lib/configs/project/service.js";
+import { fluence } from "../../helpers/commonWithSetupTests.js";
+import { MY_SERVICE_NAME, NEW_SPELL_NAME } from "../../helpers/constants.js";
+import { getServiceDirPath } from "../../helpers/paths.js";
 import {
   assertLogsAreValid,
   createSpellAndAddToDeal,
   deployDealAndWaitUntilDeployed,
-  getServiceDirPath,
+  initializeTemplate,
   updateFluenceConfigForTest,
   waitUntilShowSubnetReturnsExpected,
-} from "../sharedSteps.js";
+} from "../../helpers/sharedSteps.js";
+import { maybeConcurrentTest } from "../../helpers/testWrapper.js";
 
 describe("Deal deploy tests", () => {
   maybeConcurrentTest(
