+++ test/tests/jsToAqua.test.ts
@@ -20,8 +20,8 @@ import {
   jsToAqua,
   jsToAquaImpl,
   makeOptional,
-} from "../src/lib/helpers/jsToAqua.js";
-import { stringifyUnknown } from "../src/lib/helpers/utils.js";
+} from "../../src/lib/helpers/jsToAqua.js";
+import { stringifyUnknown } from "../../src/lib/helpers/utils.js";
 
 const fileName = "someModule";
